requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pypdf>=3.17.0
chromadb>=0.4.0

python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
